User registration process
Creation of a template with: name, email, 2x password for registration,
in the controller, the user and the verification token are created,
The controller method has the @Transactional annotation to ensure
that both the user and the and token. In the event of an error,
nothing will be created,
After successful registration, a verifikation link is sent to the email
user is sent a verification link, clicking on it changes the
in the 'user' table in the 'enabled' column to true,
The verification link expires after one day.