Create of a page for learning phrases.
In the controller, I retrieve all phrases that have a date earlier
than the current date,
The attributes of the 'Phrases' entity go into the <ul> list,
The list elements are then queued in javascript,
The again button takes the item off the queue and
inserts it at the end of the queue,
The 1day button changes the 'nextDate' of the phrase to the next day
and the 'Progress' decreases by 1 (unless it is already 1,
in which case nothing happens),
subsequent buttons increase 'progress' by 1 and change 'nextDate'
according to the formula: button value (1, 2 or 3) * progress.
The form can be submitted when the queue
 is already empty or earlier by pressing the 'finish' button.
I added an undo button to the previous page in 'groupPhrasesView' and 'phrases'.