Change in image upload functionality
The name and type of the column
in the 'group_phrases' table have been changed to 'imageUrl'.
Images are now added to Cloudinary,
and the file URL is stored in the database.
Server-side file validation has been added.
It checks whether the file has the appropriate extension (image/*)
and whether its size does not exceed 10 MB.